+++ app/model/core/src/commonMain/kotlin/com/moneymanager/domain/CsvStrategyImportExport.kt
@@ -1,4 +1,4 @@
-package com.moneymanager.domain.port
+package com.moneymanager.domain
 
 import com.moneymanager.domain.model.AppVersion
 import com.moneymanager.domain.model.csvstrategy.CsvAccountMapping
