+++ app/model/core/src/commonMain/kotlin/com/moneymanager/domain/Maintenance.kt
@@ -1,4 +1,4 @@
-package com.moneymanager.domain.port
+package com.moneymanager.domain
 
 import kotlin.time.Duration
