+++ app/ui/core/src/commonMain/kotlin/com/moneymanager/ui/AppServices.kt
@@ -3,9 +3,9 @@ package com.moneymanager.ui
 import com.moneymanager.database.Application
 import com.moneymanager.database.service.CsvStrategyExportService
 import com.moneymanager.domain.model.DeviceId
-import com.moneymanager.domain.port.CsvStrategyImportExport
-import com.moneymanager.domain.port.EntitySource
-import com.moneymanager.domain.port.Maintenance
+import com.moneymanager.domain.CsvStrategyImportExport
+import com.moneymanager.domain.EntitySource
+import com.moneymanager.domain.Maintenance
 import com.moneymanager.domain.repository.AccountAttributeRepository
 import com.moneymanager.domain.repository.AccountRepository
 import com.moneymanager.domain.repository.ApiImportStrategyRepository
