+++ app/ui/core/src/commonMain/kotlin/com/moneymanager/ui/screens/csvstrategy/CsvStrategiesScreen.kt
@@ -45,9 +45,9 @@ import com.moneymanager.domain.model.AppVersion
 import com.moneymanager.domain.model.csv.CsvImport
 import com.moneymanager.domain.model.csv.CsvRow
 import com.moneymanager.domain.model.csvstrategy.CsvImportStrategy
-import com.moneymanager.domain.port.CsvImportParseResult
-import com.moneymanager.domain.port.CsvStrategyImportExport
-import com.moneymanager.domain.port.EntitySource
+import com.moneymanager.domain.CsvImportParseResult
+import com.moneymanager.domain.CsvStrategyImportExport
+import com.moneymanager.domain.EntitySource
 import com.moneymanager.domain.repository.AccountRepository
 import com.moneymanager.domain.repository.AttributeTypeRepository
 import com.moneymanager.domain.repository.CategoryRepository
