+++ app/ui/core/src/commonMain/kotlin/com/moneymanager/ui/screens/csvstrategy/ImportStrategyDialog.kt
@@ -51,12 +51,12 @@ import com.moneymanager.domain.model.Currency
 import com.moneymanager.domain.model.csvstrategy.CsvAccountMapping
 import com.moneymanager.domain.model.csvstrategy.CsvImportStrategyId
 import com.moneymanager.domain.model.csvstrategy.export.CsvAccountMappingExport
-import com.moneymanager.domain.port.CsvImportParseResult
-import com.moneymanager.domain.port.CsvReferenceType
-import com.moneymanager.domain.port.CsvResolution
-import com.moneymanager.domain.port.CsvStrategyImportExport
-import com.moneymanager.domain.port.CsvUnresolvedReference
-import com.moneymanager.domain.port.EntitySource
+import com.moneymanager.domain.CsvImportParseResult
+import com.moneymanager.domain.CsvReferenceType
+import com.moneymanager.domain.CsvResolution
+import com.moneymanager.domain.CsvStrategyImportExport
+import com.moneymanager.domain.CsvUnresolvedReference
+import com.moneymanager.domain.EntitySource
 import com.moneymanager.domain.repository.AccountRepository
 import com.moneymanager.domain.repository.CategoryRepository
 import com.moneymanager.domain.repository.CsvAccountMappingRepository
