+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,11 +1,9 @@
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
-# Gradle version sync: keep this aligned with .github/actions/gradle-setup/action.yml
-# and .github/workflows/lint-format.yml.
-distributionUrl=https\://services.gradle.org/distributions/gradle-9.5.0-bin.zip
+distributionUrl=https\://services.gradle.org/distributions/gradle-9.5.1-bin.zip
 networkTimeout=10000
-retries=3
-retryBackOffMs=1000
+retries=0
+retryBackOffMs=500
 validateDistributionUrl=true
 zipStoreBase=GRADLE_USER_HOME
 zipStorePath=wrapper/dists
